+++ .github/actions/detect-changes/action.yml
@@ -1,0 +1,14 @@
+name: Detect Changes
+description: Determines what areas of the framework have been changed
+
+outputs:
+  docs:
+    description: If *only* docs have changed
+  rsc:
+    description: If RSC-related changes have been made
+  ssr:
+    description: If SSR-related changes have been made
+
+runs:
+  using: node20
+  main: detectChanges.mjs
