--- .github/actions/only_doc_changes/action.yml
@@ -1,8 +1,0 @@
-name: Docs
-description: Determines if the PR only changes docs
-outputs:
-  only-doc-changes:
-    description: If the PR only changes docs
-runs:
-  using: node20
-  main: only_doc_changes.mjs
