--- .github/actions/only_doc_changes/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "only_doc_changes",
-  "private": true,
-  "dependencies": {
-    "@actions/core": "1.10.0",
-    "@actions/exec": "1.1.1"
-  },
-  "packageManager": "yarn@3.6.3"
-}
